+++ backend/messages/src/client.rs
@@ -14,9 +14,9 @@ pub enum ClientMessage {
     /// Message to send Job Config
     JobConfig {
         /// Job timeout
-        timeout: usize,
+        timeout: i32,
         /// Cluster Size
-        cluster_size: usize,
+        cluster_size: i32,
         /// Types of each column in dataset for job
         column_types: Vec<String>,
         /// The name of the prediction column
