+++ backend/models/src/jobs.rs
@@ -19,9 +19,9 @@ pub struct JobConfiguration {
     /// The identifier of the dataset to be processed
     pub dataset_id: ObjectId,
     /// The timeout required for clients to process within
-    pub timeout: usize,
+    pub timeout: i32,
     /// The cluster size for a job
-    pub cluster_size: usize,
+    pub cluster_size: i32,
     /// The types of each column in the dataset
     pub column_types: Vec<String>,
     /// The column to predict during evaluation
